+++ src/control_plane/service_strategy/service_deployment_strategy.rs
@@ -8,6 +8,7 @@ pub(crate) enum ServiceDeploymentStrategy {
     DedicatedUntilIsolationProven,
     ProjectApplication,
     ProjectProcess,
+    ProjectScheduledCommand,
     Ephemeral,
 }